Visual servoing with aruco on the robot
