--- irob_subtask_logic/launch/camera.launch.autosave
@@ -1,22 +1,0 @@
-!-- Launch file for vessel cutting training excercise -->
-
-<launch>
-    <group ns="saf">
-
-
-
-        <arg name="speed_cartesian" default="10.0"/>
-        <arg name="speed_jaw" default="20.0"/>
-
-        <node name="camera" pkg="irob_subtask_logic"
-            type="camera" output="screen">
-
-                <param name="speed_cartesian" type="double" value="$(arg speed_cartesian)" />
-
-                <rosparam param="arm_names">["arm_1"]</rosparam>
-
-
-        <remap from="target" to="qrkam" />
-    </node>
-    </group>
-</launch>
