+++ irob_vision_support/launch/aruco_detector.launch
@@ -3,10 +3,12 @@
 <launch>
     <group ns="saf">
 
+    <arg name="image_input" default="stereo/left/image_rect_color"/>
+
     <node name="aruco_detector" pkg="irob_vision_support"
             type="aruco_detector.py" output="screen" >
 
-                <remap from="image_input" to="stereo/left/image_rect_color" />
+                <remap from="image_input" to="stereo/left/image_color" />
                 <remap from="image_markers" to="vision/image_markers" />
                 <remap from="markers" to="vision/markers" />
     </node>
